refactor db
